Fix other tests
Because floats are now generally bigger.

+++ fastparquet/test/test_read.py
@@ -300,5 +300,5 @@ def test_null_sizes(tempdir):
     df = pd.DataFrame({'a': [True, None], 'b': [3000, np.nan]}, dtype="O")
     fastparquet.write(tempdir, df, has_nulls=True, file_scheme='hive')
     pf = fastparquet.ParquetFile(tempdir)
-    assert pf.dtypes['a'] == 'float32'
+    assert pf.dtypes['a'] == 'float16'
     assert pf.dtypes['b'] == 'float64'
